separating logging appenders into sub mods
