+++ src/common/base/src/logging/appender/console.rs
@@ -1,0 +1,1 @@
+pub(super) struct ConsoleAppenderConfig {}
