+++ src/common/base/src/logging/mod.rs
@@ -24,7 +24,7 @@ use crate::config::placement_center::placement_center_conf;
 use crate::error::log_config::LogConfigError;
 use crate::tools::{file_exists, read_file, try_create_fold};
 
-mod config;
+mod appender;
 
 pub fn init_placement_center_log() -> Result<Vec<WorkerGuard>, LogConfigError> {
     let conf = placement_center_conf();
@@ -76,18 +76,18 @@ pub fn init_tracing_subscriber(
         }
     }
 
-    let config: config::Config = toml::from_str(&content)?;
+    let config: appender::Config = toml::from_str(&content)?;
     init_tracing_subscriber_with_config(config)
 }
 
 fn init_tracing_subscriber_with_config(
-    config: config::Config,
+    config: appender::Config,
 ) -> Result<Vec<WorkerGuard>, LogConfigError> {
     let mut layers = Vec::with_capacity(config.appenders.len());
     let mut guards = Vec::with_capacity(config.appenders.len());
 
     for (_name, appender_conf) in config.appenders {
-        let (layer, guard) = appender_conf.try_into_layer()?;
+        let (layer, guard) = appender_conf.try_into_layer_and_guard()?;
         layers.push(layer);
         if let Some(guard) = guard {
             guards.push(guard);
